Trait on_wire::FromWire

pub trait FromWire: Sized {
    fn from_bytes(wire: Vec<u8>) -> Result<Self, String>;
}

There are a lot of different ways of communicating with canisters and there's a lot of different ways of encoding those communications. Below is a non exhaustive list of the two and the Cartesian product of functions that we were creating.

| | query | inter canister call | install | endpoint | .. | bytes | query_bytes | call_bytes | .. | over_bytes | .. | json | query_json | call_json | .. | over_json | .. | candid | .. | call_canid | .. | .. | .. | proto | .. | call_proto | .. | .. | ..

The aim of this package is to abstract and disambiguate over serialization formats when deciding how to communicate with a canister. It is also designed so new communication methods/ serialization formats can easily be added to the Internet Computer.

It exists in it's own package because it's both used by canisters and the internet computers test systems.

Required methods

fn from_bytes(wire: Vec<u8>) -> Result<Self, String>

Loading content...